feat: add product y ajustes
