+++ src/app/shared/navbar.component.vue
@@ -1,8 +1,13 @@
 <script setup>
-import 'primeicons/primeicons.css'
+import { useI18n } from 'vue-i18n';
 import { useRouter } from 'vue-router';
 
+const { t, locale } = useI18n();
 const router = useRouter();
+
+const toggleLanguage = () => {
+  locale.value = locale.value === 'es' ? 'en' : 'es';
+};
 
 const navigateTo = (route) => {
   router.push(route);
@@ -15,17 +20,23 @@ const navigateTo = (route) => {
       <img src="../../assets/logo.svg" alt="StockWise" class="logo" />
     </div>
     <nav class="main-nav">
-      <a @click="navigateTo('/dashboard')" class="nav-link">Inicio</a>
-      <a @click="navigateTo('/buscar')" class="nav-link search">
-        <i class="fas fa-search"></i> Buscar
+      <a @click="navigateTo('/dashboard')" class="nav-link">{{ $t('navbar.home') }}</a>
+      <a class="nav-link search">
+        <i class="pi pi-search"></i> {{ $t('navbar.search') }}
       </a>
-      <a @click="navigateTo('/alertas')" class="nav-link">
-        <i class="fas fa-bell"></i> Alerta Stock
+      <a @click="navigateTo('/alerts')" class="nav-link alerts">
+        <i class="pi pi-bell"></i> {{ $t('navbar.alerts') }}
       </a>
-      <a @click="navigateTo('/soporte')" class="nav-link">Soporte</a>
+      <a class="nav-link">{{ $t('navbar.support') }}</a>
+
       <a @click="navigateTo('/profile')" class="nav-link user">
-        user
+        {{ $t('navbar.profile') }} <i class="pi pi-user"></i>
       </a>
+      <div class="language-switcher">
+        <button @click="toggleLanguage" class="language-button">
+          <span class="language-icon">🌐</span>
+        </button>
+      </div>
     </nav>
   </header>
 </template>
@@ -107,6 +118,31 @@ const navigateTo = (route) => {
   margin-left: 0.25rem;
 }
 
+.language-switcher {
+  margin: 0 0.5rem;
+}
+
+.language-button {
+  background: none;
+  border: none;
+  cursor: pointer;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 32px;
+  height: 32px;
+  border-radius: 50%;
+  transition: background-color 0.2s;
+}
+
+.language-button:hover {
+  background-color: rgba(0, 0, 0, 0.05);
+}
+
+.language-icon {
+  font-size: 1.1rem;
+}
+
 /* Para agregar los iconos de Font Awesome */
 @import url('https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0-beta3/css/all.min.css');
-</style> 
+</style>
